Fix compile error

+++ bot/SpendingManager.cpp
@@ -82,8 +82,8 @@ double DefaultScore (UNIT_TYPEID unitType) {
 Cost CostOfUpgrade(UpgradeID upgrade) {
     auto& data = bot.Observation()->GetUpgradeData()[upgrade];
     Cost result = {
-            data.mineral_cost,
-            data.vespene_cost,
+            (int)data.mineral_cost,
+            (int)data.vespene_cost,
             0,
             UNIT_TYPEID::INVALID
     };
